+++ packages/openai/src/responses/openai-responses-prepare-tools.ts
@@ -5,6 +5,8 @@ import {
 } from '@ai-sdk/provider';
 import { OpenAIResponsesTool } from './openai-responses-api-types';
 import { fileSearchArgsSchema } from '../tool/file-search';
+import { codeInterpreterArgsSchema } from '../tool/code-interpreter';
+import { webSearchPreviewArgsSchema } from '../tool/web-search-preview';
 
 export function prepareResponsesTools({
   tools,
@@ -22,7 +24,8 @@ export function prepareResponsesTools({
     | 'required'
     | { type: 'file_search' }
     | { type: 'web_search_preview' }
-    | { type: 'function'; name: string };
+    | { type: 'function'; name: string }
+    | { type: 'code_interpreter' };
   toolWarnings: LanguageModelV2CallWarning[];
 } {
   // when the tools array is empty, change it to undefined to prevent errors:
@@ -47,7 +50,7 @@ export function prepareResponsesTools({
           strict: strictJsonSchema,
         });
         break;
-      case 'provider-defined':
+      case 'provider-defined': {
         switch (tool.id) {
           case 'openai.file_search': {
             const args = fileSearchArgsSchema.parse(tool.args);
@@ -62,25 +65,35 @@ export function prepareResponsesTools({
             });
             break;
           }
-          case 'openai.web_search_preview':
+          case 'openai.web_search_preview': {
+            const args = webSearchPreviewArgsSchema.parse(tool.args);
             openaiTools.push({
               type: 'web_search_preview',
-              search_context_size: tool.args.searchContextSize as
-                | 'low'
-                | 'medium'
-                | 'high',
-              user_location: tool.args.userLocation as {
-                type: 'approximate';
-                city: string;
-                region: string;
-              },
+              search_context_size: args.searchContextSize,
+              user_location: args.userLocation,
             });
             break;
-          default:
+          }
+          case 'openai.code_interpreter': {
+            const args = codeInterpreterArgsSchema.parse(tool.args);
+            openaiTools.push({
+              type: 'code_interpreter',
+              container:
+                args.container == null
+                  ? { type: 'auto', file_ids: undefined }
+                  : typeof args.container === 'string'
+                    ? args.container
+                    : { type: 'auto', file_ids: args.container.fileIds },
+            });
+            break;
+          }
+          default: {
             toolWarnings.push({ type: 'unsupported-tool', tool });
             break;
+          }
         }
         break;
+      }
       default:
         toolWarnings.push({ type: 'unsupported-tool', tool });
         break;
@@ -102,11 +115,11 @@ export function prepareResponsesTools({
       return {
         tools: openaiTools,
         toolChoice:
-          toolChoice.toolName === 'file_search'
-            ? { type: 'file_search' }
-            : toolChoice.toolName === 'web_search_preview'
-              ? { type: 'web_search_preview' }
-              : { type: 'function', name: toolChoice.toolName },
+          toolChoice.toolName === 'code_interpreter' ||
+          toolChoice.toolName === 'file_search' ||
+          toolChoice.toolName === 'web_search_preview'
+            ? { type: toolChoice.toolName }
+            : { type: 'function', name: toolChoice.toolName },
         toolWarnings,
       };
     default: {