+++ packages/llamaindex/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/llamaindex
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [63a5dc5]
+  - ai@5.0.10
+
 ## 1.0.9
 
 ### Patch Changes